requirements_airflow.txt
pandas==2.0.3
requests==2.31.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.16.0
rdkit==2023.3.1